test: test fixes visibility for ops contr role
