+++ tests/app/Filament/IntentionCodeResourceTest.php
@@ -16,7 +16,7 @@ use Livewire\Livewire;
 class IntentionCodeResourceTest extends BaseFilamentTestCase
 {
     use ChecksDefaultFilamentAccess;
-    use ChecksFilamentActionVisibility;
+    use ChecksDefaultFilamentActionVisibility;
 
     public function testItLoadsAnIntentionCodeForViewWithAirfieldIdentifier()
     {
