+++ tests/app/Filament/PrenoteResourceTest.php
@@ -16,7 +16,7 @@ use Livewire\Livewire;
 class PrenoteResourceTest extends BaseFilamentTestCase
 {
     use ChecksDefaultFilamentAccess;
-    use ChecksFilamentActionVisibility;
+    use ChecksDefaultFilamentActionVisibility;
 
     public function testItLoadsDataForView()
     {
